+++ components/BottomWiki/index.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import SimpxLogo from '../../assets/page/simpx.png';
 
 import './bottomwiki.css';
 
@@ -7,7 +6,7 @@ export default class BottomWiki extends React.Component{
 
   render(){
     return(
-        <img id='bottomwiki' src={SimpxLogo}/>
+        <div id='bottomwiki'/>
       );
   }
 }