+++ pages/index.js
@@ -1,8 +1,12 @@
 import React from 'react';
-import Link from 'next/link';
-import fetch from 'isomorphic-unfetch'
+import fetch from 'isomorphic-unfetch';
 import Head from 'next/head';
+import { IconButton } from '@material-ui/core';
+import { Search } from '@material-ui/icons';
+
 import userConfig from '../services/UserConfig';
+import { ArticleList } from '../components';
+import './index.css';
 
 export default class Index extends React.Component{
     state = {
@@ -23,49 +27,32 @@ export default class Index extends React.Component{
         res = await fetch('http://'+window.location.hostname+':5000/articles') :
         res = await fetch('http://'+window.location.hostname+'/articles');
 
-        //const res = await fetch('http://'+window.location.hostname+'/articles');
         let data = await res.json();
 
         this.setState({os, lang, arch, data});
     }
 
     render(){
-        const { data } = this.state;
-
         return(
             <>
                 <Head>
                     <title>Dyn-Wiki - A dynamic Wiki for easy reading</title>
                 </Head>
-                <div>
-                    <h1>Dyn-Wiki - A dynamic Wiki for easy reading</h1>
-                    Dados pelo Navegador:<br/>
-                    Sistema Operacional: {this.state.os}<br/>
-                    Idioma: {this.state.lang}<br/>
-                    Arquitetura: {this.state.arch}
-                    <br/><br/>
-                    Artigos:
-                    <ul>
-                    {
-                        data.length !== 0 ? data.map(article =>{
-                            let res = article.article;
-                            return <li>{res}<ul>{article.versions.map(version => {
-                                let res2 = version.version;
-                                return <li>{res2}<ul>{version.langs.map(lang => {
-                                    return(
-                                        <Link href={`/wiki/${res}/${res2}/${lang.abr}`}>
-                                            <a>
-                                                <li>{lang.title} [{lang.abr}]<br/>
-                                                    {lang.desc}
-                                                </li>
-                                            </a>
-                                        </Link>
-                                    );
-                                })}</ul></li>
-                            })}</ul></li>
-                        })
-                    : null } 
-                    </ul>
+                <div className='root-container'>
+                    <div className='top-container'>
+                        <span className='top-dynwiki'>Dyn-Wiki</span>
+                        <div className='top-input'>
+                            <div className='input-container'>
+                                <input type='text' className='top-input-field'/>
+                            </div>
+                            <IconButton>
+                                <Search className='input-search'/>
+                            </IconButton>
+                        </div>
+                    </div>
+                    <div className='bottom-container'>
+                        { <ArticleList/> }
+                    </div>
                 </div>
             </>
         );