+++ services/Articles/Articles.js
@@ -3,20 +3,30 @@ const fs = require('fs');
 
 /**
  * @class Article
+ * @author vinicius-portela
+ * @version 0.2.0
+ * 
  * Manipulate article info, directory, etc
  */
 class Article {
-    cmd = {
-        verbose: false,
-        //Default Configuration
-        defaultConfig: {
-            arch: 'x64',
-            os: 'linux',
-            os_version: '5.0.0-31-generic',
-            dist: 'kde_neon',
-            dist_version: '5.16',
-            swap: 10843,
-            swap_type: 'partition',
+
+    /**
+     * Default Command Configurations
+     * @returns Default Configuration
+     */
+    static defaultCmd(){
+        return {
+            verbose: false,
+            // Default Configuration, It's temporary, Just for Testing
+            defaultConfig: {
+                arch: 'x64',
+                os: 'linux',
+                os_version: '5.0.0-31-generic',
+                dist: 'kde_neon',
+                dist_version: '5.16',
+                swap: 10843,
+                swap_type: 'partition',
+            }
         }
     }
 
@@ -25,7 +35,7 @@ class Article {
      * @param {String} folder - Path to Folder 
      * @return {Array} - List of Versions
      */
-    getVersions(folder){
+    static getVersions(folder){
         let res = [];
 
         let versions = fs.readdirSync(folder);
@@ -50,7 +60,7 @@ class Article {
      * @param {String} folder - Path to Folder
      * @return {Array} - List of Languages
      */
-    getLangs(folder){
+    static getLangs(folder){
         let res = [];
 
         let langs = fs.readdirSync(folder);
@@ -64,10 +74,12 @@ class Article {
 
     /**
      * Get all articles inside a folder
+     * @param {String} folder - folder name from get Articles
+     * @param {Object} cmd - cmd configuration
      * @return Array with all Articles
      */
-    getArticles(folder){
-        const { verbose } = this.cmd;
+    static getArticles(folder, cmd=this.defaultCmd()){
+        const { verbose } = cmd;
         if(verbose) console.log(`[V] Getting Articles of folder '${folder}'`);
         
         let res = [];
@@ -100,12 +112,19 @@ class Article {
 
                 //Separe Languages
                 langs.forEach(lang => {
+                    const file = folder+article+'/'+version+'/'+lang;
                     let curLang = /(?:[.])(.*?)(?:[.])/gm.exec(lang);
 
-                    let title = reader.config(folder+article+'/'+version+'/'+lang, 'title');
-                    let desc = reader.config(folder+article+'/'+version+'/'+lang, 'desc');
+                    let title = reader.config(file, 'title');
+                    let desc = reader.config(file, 'desc');
+                    let img = reader.config(file, 'article_image');
 
-                    res[article_index].versions[version_index].langs.push({abr:curLang[1],title, desc});
+                    res[article_index].versions[version_index].langs.push({
+                        abr:curLang[1],
+                        title, 
+                        desc,
+                        img,
+                    });
                 });
             });
         });
@@ -114,20 +133,14 @@ class Article {
     }
 
     /**
-     * Check for all tags inside a Article, this is used for getting all possible
-     * configurations to be saved in cache
-     */
-    checkForTags(){}
-
-    /**
      * Check if a article is saved in cache
      * @param {String} article - The Article ID
      * @param {String} version - The Article Version
      * @param {String} lang - The Article Language (PT, EN, ES, JP, DE ...)
      * @returns - True/False Result
      */
-    isInCache(article, version, lang){
-        const inCache = fs.existsSync(`services/Articles/Articles/${article}/${version}/${lang}.json`);
+    static isInCache(article, version, lang){
+        const inCache = fs.existsSync(`services/Articles/cache/${article}/${version}/${lang}.json`);
 
         if(inCache) return true; else return false;
     }
@@ -136,14 +149,14 @@ class Article {
      * Save a file in cache
      * @param {Object} data - The Result Object with all Data
      */
-    toCache(data){
+    static toCache(data){
         //Get Article Metadata
         const { article, version, lang } = data.meta;
 
         const content = JSON.stringify(data);
-        const baseDir = 'services/Articles/Articles'
+        const baseDir = 'services/Articles/cache'
 
-        //Create Dir
+        //Create Dir if doesn't exists
         const hasArticleRoot = fs.existsSync(`${baseDir}`);
         const hasArticle = fs.existsSync(`${baseDir}/${article}`);
         const hasVersion = fs.existsSync(`${baseDir}/${article}/${version}`);
@@ -156,33 +169,36 @@ class Article {
             console.log('💾 Saved to cache')
         });
     }
-    
-    //TODO: Has to import config too, so see if a article is in cache or not
-    // Because the user can have a different config that doesnt have an article in cache yet
 
     /**
      * Loads a file from cache
      * @param {String} article - The Article ID
      * @param {String} version - The Article Version
      * @param {String} lang - The Article Language (PT, EN, ES, JP, DE ...)
+     * @param {Object} config - The User Configuration
+     * @returns {Object} Object prepared to PageRender
      */
-    loadFromCache(article, version, lang){
-        const res = fs.readFileSync(`./services/Articles/Articles/${article}/${version}/${lang}.json`);
-        return res;
+    static loadFromCache(article, version, lang, config){
+        let data = JSON.parse(fs.readFileSync(`./services/Articles/cache/${article}/${version}/${lang}.json`, 'utf-8'));
+        
+        return this.justUserConfig(data, config)
     }
 
     /**
-     * It's part of process of get(), it's more background service than get()
+     * It's part of process of get(), it's more internal service than get()
      * @param {String} article - The Article ID
      * @param {String} version - The Article Version
      * @param {String} lang - The Article Language (PT, EN, ES, JP, DE ...)
+     * @param {Object} config - The user Configuration
+     * @returns Object with all data for PageRender
+     * @internal
      */
-    convert(article, version, lang){
+    static convert(article, version, lang, config){
         let file = `articles/${article}/${version}/article.${lang}.md`
-        let userConfig = require('../UserConfig/userConfig.example');
+        //let userConfig = require('../UserConfig/userConfig.example');
 
         //First Load File (With MDReader)
-        let data = reader.convert(file, userConfig);
+        let data = reader.convert(file);
 
         //Add Metadata
         data.meta = {
@@ -198,7 +214,7 @@ class Article {
         //Finally Add to Cache for Next Searchs
         this.toCache(data);
 
-        return data;
+        return this.justUserConfig(data, config);
     }
 
     /**
@@ -209,20 +225,19 @@ class Article {
      * @param {String} config - The User System Configuration
      * @returns - The Article in JSON
      */
-    get(article, version, lang, config = this.defaultConfig){
+    static get(article, version, lang, config = this.defaultCmd().defaultConfig){
         const hasInCache = this.isInCache(article, version, lang);
-        //const hasInCache = false;
 
         //Check if file already in cache
         if(hasInCache){
             //In cache
-            let data = this.loadFromCache(article, version, lang);
+            let data = this.loadFromCache(article, version, lang, config);
             console.log('📀 Already in Cache, Loading...');
 
             return data;
         }else{
             //Not in Cache
-            let data = this.convert(article, version, lang);
+            let data = this.convert(article, version, lang, config);
 
             return data;
         }
@@ -232,7 +247,7 @@ class Article {
      * Update list of articles
      * Used by the Express Node Server
      */
-    update(){
+    static update(){
         console.log('🔄 Updating Article List...');
         let updated_data = this.getArticles('./articles/');
         let updated_content = JSON.stringify(updated_data);
@@ -240,6 +255,93 @@ class Article {
         if(err) throw error;
         });
     }
+
+    /**
+     * Check if a tag is compatible with user configuration
+     * @param {String} tag 
+     * @param {Object} tagsConfig 
+     */
+    static checkCompatibility(tag, tagsConfig, userConfig=this.defaultCmd().defaultConfig){
+        //Check if tag exists in tagsConfig
+        if(tagsConfig.hasOwnProperty(tag)){
+            //For each needed configuration in tagsConfig
+            let keys =  Object.keys(tagsConfig[tag])
+            
+            for(let i=0; i<keys.length ; i++){
+                let item = keys[i]
+                let user = userConfig[item]
+                if (!tagsConfig[tag][item].includes(user))
+                    return false
+            }
+
+            return true
+        }
+    }
+
+    /**
+     * Convert files that aren't in articles/ file structure
+     * @param {Array} arr - Array with all Article Content
+     * @param {Object} config - Configuration of User
+     */
+    static justUserConfig(arr, config=this.defaultCmd().defaultConfig){
+        //Get configurations of tags
+        const tags = require('../../tags.config')
+
+        let array = arr.data
+
+        //Get all unique scopes from array
+        let scopes = []
+
+        array.map(item => {
+            if(item.hasOwnProperty('scope'))
+                if(scopes.indexOf(item.scope) === -1)
+                    scopes.push(item.scope)
+        })
+
+        //Check for compatibility
+        scopes.forEach(tag => {
+            let comp = this.checkCompatibility(tag, tags, config)
+
+            if(comp){
+                //Remove from scope and insert
+
+                //Remove from scope
+                for(let i=0 ;  i<array.length ; i++){
+                    let item = array[i]
+                    if(item.hasOwnProperty('scope')){
+                        if(item.scope === tag){
+                            let content = []
+                            item.data.forEach(item => content.push(item))
+                            
+                            array = array.slice(0, i).concat(content, array.slice(i+1))
+                        }
+                    }
+                }
+                
+            }else{
+                //Just Remove
+                array = array.filter(item => {
+                    if(!item.hasOwnProperty('scope')){
+                        return true
+                    }else{
+                        if(item.scope !== tag)
+                            return true
+                    } 
+                })
+            }
+        })
+
+        arr.data = array
+
+        return arr;
+
+        //let content = 'module.exports = ' + JSON.stringify(arr)
+        //fs.writeFileSync('./data.js', content)
+    }
 }
 
-module.exports = new Article;
+/* FOR DEBUG */
+//Article.justUserConfig(require('../MDReader/data'))
+/* FOR DEBUG */
+
+module.exports = Article;